+++ src/components/location/LocationDetail.js
@@ -10,12 +10,17 @@ const LocationDetail = props => {
     const [isLoading, setIsLoading] = useState(true);
 
     const handleLocationDelete = () => {
-        if (window.confirm("Are you sure you want to close this location?")) {
-            setIsLoading(true);
-            LocationManager.delete(props.locationId).then(() =>
-                props.history.push("/locations")
-            );
+        if (employees.length === 0) {
+            if (window.confirm("Are you sure you want to close this location?")) {
+                setIsLoading(true);
+                LocationManager.delete(props.locationId).then(() =>
+                    props.history.push("/locations")
+                );
+            }
+        } else {
+            window.alert("Please reassign all employees working at this location before closing the location.")
         }
+        
     };
 
     const handleEmployeeDelete = (employeeId) => {
